Greatly consolidates visibilitySwitch using util.js, if the engine doesnt allow for imports, all my cute clean ups will be for naught
